refactor(MapGenerator-Python): :recycle: adjust code structure
