+++ dependency/MapGenerator-Python/Classes/AreaRenderDict.py
